Both Yakuake and KRunner used to suffer from the following bug: If there are two monitors, the top edges of which are alined, and on one of those monitors (but not the other) is a panel at the top edge, their windows would appear at a distance away from the top edge equal in height to the panel when appearing on the monitor without a panel.
This change fixes this bug in Yakuake generating a list of windows which
have top edge struts intersecting the geometry of the current screen, and
then handing that list to the variant of KWindowSystem::workArea() that
takes a list of windows to ignore in computing the work area as parameter.

The same change ought to be applied to KRunner, or a new workArea()
variant that does this should be added to KWindowSystem for convenience.

Thanks to Lubos for useful pointers.

BUG:147601
CCMAIL:plasma-devel@kde.org
CCMAIL:llunak@kde.org

svn path=/trunk/extragear/utils/yakuake/; revision=1130207