Pull the window to the current virtual desktop under certain conditions.
When the window is set not to show on all virtual desktops and is not
residing on the current virtual desktop, but is open (i.e. set to stay
open regardless of focus loss), the Open/Retract action normally used
to either retract it, or, when the option to also use the action to
focus the window is enabled, cause the window manager to switch to the
virtual desktop the window resides on. The former case, i.e. when the
option to also use the action to focus the window is disabled, has now
been changed to move the window to the current virtual desktop and
focus it, rather than retract it.

Also augments the What's This blurb on the focus option to mention the
window manager's VD switch behavior, and added some code comments to
help charting the codepaths.

BUG:210880