Redo version checking approach for activity/silence monitoring.
Due to moc's limitations we end up having non-useful slots in the
D-Bus interface on <4.7.1. Unfortunately it's not possible to
toggle the scriptable flag on a QMetaMethod at runtime, and
QDBusConnection::registerObject() can't deal with an OR'd combo
of ExportScriptableSlots and ExportAdaptors, so it's not possible
to stuff the extra slots into a QDBusAbstractAdaptor subclass
either. I stopped short of duplicating the entire object inter-
face plus extra slots in an adaptor to toggle between the two at
runtime - as much as I hate sloppy D-Bus interfaces, that would
just be too ugly in terms of code. Maybe I'll find an acceptable
way later.