Fix geometry in multi-screen cases
Commit 81484067 addressed and issue with multi-head setups where the
top strut of a window not visible on yakuake's display would be removed
from its work area. However, it did not consider bottom, right, and left
struts, and also prevented the left, right and bottom struts of windows
on the same display as yakuake from being removed from its work area.
This patch fixes this issue and tries to make the code more explicit.

BUG: 265818