Fix scaling of the GHNS icon in the skin list delegate.
It's now basically qMin(16, text height) so the icon doesn't overlay
the ascenders of the author name if the skin name bounding rect height
is is smaller than 16 px. The downside is the blurryness of scaling,
but then with the standard font size and the trend toward increasing
screen dpi this is essentially a corner case anyway.