Turn off automatic quit when all windows close.
With QApplication::quitOnLastWindowClosed set to true (which is the
default), Yakuake would quit when for example the Configure dialog was
closed while the main window was hidden.

So we turn that off, and instead call QApplication::quit() explicitly
instead of just closing the window when the user clicks the quit button.

Acked by Eike Hein on IRC.

BUG: 345116