Force platform xcb if built with X11 support
Yakuake is at the moment completely broken on platform Wayland
(window doesn't show), but as an X11 application on Xwayland it
works just fine. So for the time being better enforce xcb platform.

If one wants to test on wayland it's still possible to start it
through the dedicated command line switch --platform wayland which
overrides the environment variable.

REVIEW: 125824