Manual set of tab title when automatic titles enabled
When "Show title bar contents in tab labels" is checked in settings, one cannot
set intectively the tab title. This patch allows to interactively set the tab
title, which is kept between change of settings. To clear to title set this
way, one just needs to edit and delete the title.