Make yakuake work on Wayland
Summary:
With this change yakuake works on Wayland. So far the xcb platform
was enforced as the yakuake window didn't show. The main reason for
this was that getDesktopGeometry() returns a useless geometry on
Wayland as there is no concept of a work area and also no way to get
the window geometries. Thus the method returns the screenGeometry
just like in full screen case.

In addition the xshape variant to toggle window state is not invoked.
Instead it's just the shared hide and show.

This commit message was typed in yakuake on Wayland.

Reviewers: hein

Differential Revision: https://phabricator.kde.org/D1091