Fix Skin validation when installed from KNS
Summary:
When a Skin is downloaded from KNS, Yakuake fails to validate the presence of title.skin and tabs.skin files.
The skin is then uninstalled even if these files exist.

This is due to this optimization of KNS: https://phabricator.kde.org/D6104
Files are not listed anymore by entry.installedFiles().
They are replaced by paths that end with /*

In order to fix this bug I changed the validateSkin method so that it checks the presence of files in the filesystem instead of in a list of entries returned by KNS.

BUG: 395012

Test Plan:
Download a Skin from KNS:  It should not complain about missing files.
Install a correct skin from local: it should not  complain about missing files;
Install a wrong skin from local: it should  complain about missing files;
Install a correct skin already installed from local: it should ask for overwrite existing skin.

Reviewers: hein, apol, #yakuake

Reviewed By: hein

Subscribers: Zren, hein, apol, chauvin

Tags: #yakuake

Differential Revision: https://phabricator.kde.org/D13335