Fix yakuake not opening on multimonitor setups
Fixes yakuake not opening on one monitor when the panel is located on an edge
between two monitors.

This patch corrects the ExtendedStruct calculations and then checks if
not only the strut intersects with the current monitor but also the
window frame. If the strut intersects but the window frame does not that
window is most likely some king of panel located on an edge between the current
and the next screen. With this patch those struts and windows will now be
ignored.