Fix shortcut conflict for session closing
The Konsole part and Yakuake both provide a close-session action with the same shortcut (Ctrl-Shift-W). This leads to a conflict when invoking it. To work around this this patch removes the shortcut from the action provided by the konsole part.

I'm not particularly happy with this patch. Ideally we'd just reuse the action provided by the part, but this seems hard to do since the part actions appear to be per session and the yakuake action is global

BUG: 319172