Fix crash when configured screen is missing
When Yakuake is configured to be on screen n and only n-1 screens are available it crashes in mainwindow.cpp:1139. To prevent this check if enough screens are available and if not return the last screen in the list.

The screen setting is not altered so once the screen becomes available again it is used.

I'm not sure about the implications of this for the screen selection menu when having > 2 screens as I don't have such a setup.