Make it possible to configure the konsolepart shortcuts
This patch allows the user to configure the by konsolepart provided actions. For example it wasn't possible to configure the copy/paste shortcuts. Also many the "X is an ambiguous shortcut" can now be fixed and remapped if the user so desires.

Internally this is done by collecting the actioncollections from the kparts and then providing the first one to the shortcutconfigure dialog. After the shortcuts have been configured the actioncollections from all the other terminals get synchronized.

BUG: 412496