Add ability to hide titlebar
# This MR adds the option to hide the title bar to provide a seamless visual effect.

![Screenshot_20201116_112104](https://invent.kde.org/utilities/yakuake/uploads/f248b52cfee904f8327d9e5f373a89c0/Screenshot_20201116_112104.png)

## Changes

To facilitate this change the borders also got an option to be hidden:
*Note:* The bottom border also moved from the skins background to `MainWindow::paintEvent`. The color and width are following the specification provided by the skin.

![image](https://invent.kde.org/utilities/yakuake/uploads/267141465e082ff35fbb3c0fc7e8a6f0/image.png)

## Safeguards

To warn the user that they are about to hide the title bar, which is (when no tray icon is displayed) the only option to access the settings (besides from to the user maybe not known keyboard shortcut) and re-show the title bar, a dialog opens when the user presses the shortcut (default: Ctrl+Shift+m following konsole's model) or is about change the setting in the settings window (red circle): ![image](https://invent.kde.org/utilities/yakuake/uploads/2e57d8eea7e5b3fc28cd64c25a219332/image.png) *Note: in this screen shot a border (yellow circles) at the bottom of the terminal is still present, this is because the skin gets loaded from the systems installation and not the development directory. This border is not part of the skin anymore. Also note that the windows on my desktop have no borders, this is a personal choice and does not affect other users.*

Choosing to not display on of the dialog anymore (checking the "Don't show this message again" checkbox) hides the dialog in both places. If the dialog is triggered by the keyboard shortcut, rejecting the dialog will keep the title bar visible. If the dialog is triggered by checking the checkbox in the settings window and the user rejects the dialog the checkbox is reset to unchecked.