Wayland: don't block on a response from plasmashell before toggling
When the system is under load, Plasma may take a while to respond, and
Yakuake is left feeling unresponsive as it's waiting on a response from
Plasma before responding to user input.

Since the available desktop region doesn't change much, we can assume that
the old value is okiedokie and go ahead and show the window according to it.

When the latest value comes in, we update the position of the window.
For the case when it's the same, nothing happens.
For the case when it's different, it moves to the correct location.