Fix Slide effect on Wayland
Qt Wayland windows have their internal Wayland surface destroyed when they're
hidden, and re-created when they're shown. Since `KWindowEffects::slideWindow()`
requires the Wayland surface, it doesn't work when called after `hide()` and
before `show()`. The solution is simply to call it directly after `show()`; window
opening animations are asynchronous so the effect is applied to the window as
it's opening.

BUG: 447456
FIXED-IN: 22.04