Force-exit from the fullscreen mode unless asked to keep the state
The expected default behaviour is for yakuake to re-open itself in
windowed mode after being closed while in fullscreen mode, unless the
"Remember and restore window fullscreen state" is set in the settings.
In this case, we should re-open the window in fullscreen mode.

On X11, we did not need to call `setFullscreen(false)` when closing
the window, as the `changeEvent()` callback would get called when
opening the window again would reset the state to the expected value.

However, on Wayland, the Window gets re-created being fullscreen, leading
to a state inconsistency.

Let's fix this by always setting the expected size *before* opening the
window.

BUG: 451684