Fix not retracting when opening config window on Wayland
Usually when the main window loses focus we retract Yakuake.
However, we don't want that to happen when opening one of our configuration dialogs.

We do that by checking whether the newly focused window has the main window as its parent.
The current code for this uses KWindowInfo and  only works on X11, on Wayland it's broken.

By using QWindow API we can make the code work on Wayland too.