org.kde.yakuake.appdata.xml add donation URL and launchable
Fix "appstreamcli validate" issue "I: org.kde.yakuake.desktop:~:
desktop-app-launchable-omitted
This `desktop-application` component has no `desktop-id` launchable tag,
however it contains all
the necessary information to display the application. The omission of the
launchable entry means
that this application can not be launched directly from installers or
software centers. If this
is intended, this information can be ignored, otherwise it is strongly
recommended to add a
launchable tag as well.".