mainwindow: Add a "Primary Screen" option
The screen number of the selected primary screen can change between
sessions.
The screen numbers can also vary, especially with laptops moving from
a docking station to another.

To have a more stable behavior, add the option to always drop down on
the selected Pimary screen (a.k.a the screen with highest priority).

This adds a dependency on libkscreen to retrieve the outputs priority
values.

Signed-off-by: Detlev Casanova <dc@detlev.ca>